+++ src/Http/RateLimitPolicySet.php
@@ -183,6 +183,7 @@ final readonly class RateLimitPolicySet
             'message_report' => self::definition(10, 3_600, 1, 100, 60, 86_400),
             'moderation_queue' => self::definition(120, 60, 1, 5_000, 1, 3_600),
             'moderation_action' => self::definition(60, 3_600, 1, 1_000, 60, 86_400),
+            'push_subscription_management' => self::definition(20, 3_600, 1, 100, 60, 86_400),
         ];
     }